# reshapr-public-openapi-v0.1.yaml
...
openapi: 3.0.2
info:
  title: reShapr Public API v0.1
  version: 0.0.1
  description: reShapr control plane REST API
  contact:
    name: Laurent Broudoux
    url: https://github.com/lbroudoux
    email: laurent.broudoux@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /v1/expositions:
    get:
      parameters:
      - name: serviceId
        description: Unique identifier of the Service to get Expositions for
        schema:
          type: string
        in: query
      - name: gatewayGroupId
        description: 'Identifier of the Gateway Group to get Expositions for '
        schema:
          type: string
        in: query
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exposition'
          description: Get the expositions - either all or filtered by Service or
            Gateway Group
      security:
      - jwt-bearer: []
      operationId: getExpositions
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpositionReference'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exposition'
          description: Exposition successfully created
        "404":
          content:
            application/json:
              schema:
                type: string
          description: One of exposition dependency (Configuration Plan or Gateway
            Group) cannot be found
      security:
      - jwt-bearer: []
      operationId: createExposition
  /v1/expositions/{id}:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exposition'
          description: The Exposition with provided id
        "404":
          description: Exposition with this id is unknown
      security:
        - jwt-bearer: [ ]
      operationId: getExposition
    delete:
      responses:
        "204":
          description: Exposition deleted
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Exposition with this is is not found
      security:
        - jwt-bearer: [ ]
      operationId: deleteExposition
    parameters:
      - name: id
        description: Exposition unique identifier
        schema:
          type: string
        in: path
        required: true
  /v1/expositions/active:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActiveExposition'
          description: The list of active Expositions
      security:
        - jwt-bearer: []
      operationId: getActiveExpositions
  /v1/expositions/active/{id}:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveExposition'
          description: The active exposition with provided id
        "404":
          description: Exposition with this id is unknown
      security:
        - jwt-bearer: [ ]
      operationId: getActiveExposition
    parameters:
      - name: id
        description: The unique identifier of the active Exposition to manage
        schema:
          type: string
        in: path
        required: true
...
components:
  schemas:
...
    ConfigurationPlanSummary:
      description: The summary of a ConfigurationPlan in reShapr control plane
      required:
      - id
      - backendEndpoint
      type: object
      properties:
        id:
          description: Unique identifier of the ConfigurationPlan
          type: string
        backendEndpoint:
          description: The backend endpoint of this Configuration Plan
          type: string
...
    ServiceView:
      description: An aggregated view no Service and related elements in reShapr
      required:
      - id
      - organizationId
      - name
      - version
      - createdOn
      - type
      type: object
      properties:
        id:
          description: Service unique identifier
          type: string
        organizationId:
          description: Identifier of the owning organization
          type: string
        name:
          description: Service name
          type: string
        version:
          description: Service version
          type: string
        createdOn:
          format: date-time
          description: Creation date in control plane
          type: string
        type:
          $ref: '#/components/schemas/ServiceType'
          description: Service type
        operations:
          description: The list of operations exposed by this service
          type: array
          items:
            $ref: '#/components/schemas/Operation'
...
    ServiceType:
      description: The available types of Service in reShapr
      enum:
      - REST
      - GRAPHQL
      - GRPC
      type: string
...
    Exposition:
      description: An Exposition in reShapr control plane
      required:
      - id
      - organizationId
      type: object
      properties:
        id:
          description: Unique identifier of the Exposition
          type: string
        organizationId:
          description: Unique identifier of the organization owning the Exposition
          type: string
        name:
          description: Optional organization-unique name of the Exposition (absent when unnamed)
          type: string
        createdOn:
          format: date-time
          description: Creation date in control plane
          type: string
        gatewayGroup:
          $ref: '#/components/schemas/GatewayGroup'
          description: The gateway group the service is exposed on
        configurationPlan:
          $ref: '#/components/schemas/ConfigurationPlan'
          description: The configuration plan used for this exposition
        service:
          $ref: '#/components/schemas/Service'
          description: The service being exposed
    ExpositionReference:
      description: A reference to an Exposition in reShapr control plane
      required:
      - id
      - organizationId
      type: object
      properties:
        id:
          description: Unique identifier of the Exposition
          type: string
        organizationId:
          description: Unique identifier of the organization owning the Exposition
          type: string
        name:
          description: Optional organization-unique name of the Exposition. When omitted at creation,
            the Exposition is created unnamed (a default slug may be proposed by the CLI/Web UI).
          type: string
        createdOn:
          format: date-time
          description: Creation date in control plane
          type: string
        gatewayGroupId:
          description: Identifier of the gateway group the service is exposed on
          type: string
        configurationPlanId:
          description: Identifier of the configuration plan used for this exposition
          type: string
    Operation:
      description: A Service Operation in reShapr control plane
      required:
      - name
      type: object
      properties:
        name:
          description: Name of the operation
          type: string
        method:
          description: HTTP or protocol specification method if it makes sense
          type: string
        action:
          description: Denoted operation action (depending on the protocol)
          type: string
        intputName:
          description: Name of input message type
          type: string
        outputName:
          description: Name of output message type
          type: string
    Service:
      description: A Service in Micepe control plane
      required:
      - id
      - organizationId
      - name
      - version
      - createdOn
      - type
      type: object
      properties:
        id:
          description: Service unique identifier
          type: string
        organizationId:
          description: Identifier of the owning organization
          type: string
        name:
          description: Service name
          type: string
        version:
          description: Service name
          type: string
        type:
          $ref: '#/components/schemas/ServiceType'
          description: Service type
        createdOn:
          format: date-time
          description: Creation date in control plance
          type: string
    ConfigurationPlan:
      description: A ConfigurationPlan in Micepe control plane
      required:
      - id
      - organizationId
      - backendEndpoint
      - name
      - serviceId
      type: object
      properties:
        id:
          description: Unique identifier of the ConfigurationPlan
          type: string
        organizationId:
          description: Identifier of the owning organization
          type: string
        backendEndpoint:
          description: The backend endpoint of this Configuration Plan
          type: string
        name:
          description: Name of the Configuration Plan
          type: string
        serviceId:
          description: Identifier of the Service this Configuration Plan relates to
          type: string
        excludedOperations:
          description: List of excluded operation names
          type: array
          items:
            type: string
        includedOperatoions:
          description: List of included operation names
          type: array
          items:
            type: string
        includedArtifacts:
          description: List of selected attached artifact names. Empty (or absent) means all the
            attached artifacts of the service apply. The service main artifact is never impacted.
          type: array
          items:
            type: string
        backendSecretId:
          description: Identifier of Backend Secret to use if any
          type: string
        apiKey:
          description: The MCP endpoint API Key if configured
          type: string
        initialAccessToken:
          description: Initial Access Token if using th einternal identity provider
            for OAuth2 DCR
          type: string
        backendTimeout:
          description: Timeout in milliseconds for requests to the backend endpoint
          type: integer
        audit:
          description: Whether the audit log will be enabled and calls logged. Default is false.
          type: boolean
    ActiveExposition:
      description: An Exposition that is currently active. ie. exposed by at least
        one Gateway
      required:
      - service
      - configurationPlan
      - gateways
      - id
      - organizationId
      type: object
      properties:
        service:
          $ref: '#/components/schemas/ServiceView'
          description: ""
        configurationPlan:
          $ref: '#/components/schemas/ConfigurationPlanSummary'
          description: ""
        gateways:
          description: ""
          type: array
          items:
            $ref: '#/components/schemas/GatewayView'
        id:
          description: Unique identifier of the Exposition
          type: string
        organizationId:
          description: Unique identifier of the organization of the Exposition
          type: string
        name:
          description: Optional organization-unique name of the Exposition (absent when unnamed)
          type: string
        createdOn:
          format: date-time
          description: Creation date in control plane
          type: string
...
    GatewayGroup:
      description: A GatewayGroup in Reshapr control plane
      required:
      - id
      - organizationId
      - name
      type: object
      properties:
        id:
          description: The unique identifier of this GatewayGroup
          type: string
        organizationId:
          description: The unique identifier of the organization this group is attahced
            to
          type: string
        name:
          description: The unique name of this GatewayGroup
          type: string
        labels:
          description: The labels attached to this gateway group (key=value pairs)
          type: object
          additionalProperties: true
    GatewayView:
      description: A gateway currently active and reporting health to the Reshapr control plane
      required:
      - id
      - name
      type: object
      properties:
        id:
          description: The unique identifier of this Gateway
          type: string
        organizationId:
          description: Identifier of the organization owning this gateway
          type: string
        name:
          description: The unique name of this Gateway
          type: string
        version:
          description: The version of the gateway runtime advertised during registration
          type: string
        startedAt:
          format: date-time
          description: The timestamp when this gateway was started
          type: string
        lastHeartbeat:
          format: date-time
          description: The timestamp of the last heartbeat received from this gateway
          type: string
        fqdns:
          description: The list of fully-qualified domain names served by this gateway
          type: array
          items:
            type: string
        labels:
          description: The labels advertised by this gateway during registration (key=value pairs)
          type: object
          additionalProperties: true
        gatewayGroups:
          description: The gateway groups this gateway is mapped to
          type: array
          items:
            $ref: '#/components/schemas/GatewayGroup'
...
  securitySchemes:
    jwt-bearer:
      scheme: bearer
      type: http
      description: JWT Bearer acquired from reShapr Authentication API /auth endpoint